pyproject.toml
[project]
name = "geekmagic-hacs"
version = "1.0.0"
description = "Home Assistant custom integration for GeekMagic displays"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Adrien Brault" }
]
dependencies = [
    "pillow>=10.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-cov>=4.0.0",
    "homeassistant>=2024.1.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["custom_components/geekmagic"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "homeassistant>=2024.1.0",
    "voluptuous>=0.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.coverage.run]
source = ["custom_components/geekmagic"]
